requirements.txt
Django==5.1.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
psycopg2-binary==2.9.*
Pillow==11.0.*
python-decouple==3.8
python-dotenv==1.0.*
black==24.*
isort==5.*
pre-commit==3.*

